[01295] Migrate DataTable ghost variant sample to DataTableApp tabs
- Removed DataTableGhostApp.cs
- Added DataTableGhostSample to DataTableApp.cs with side-by-side comparison in cards
- Added dark mode fix for ghost variant horizontal borders in useTableTheme.ts

Made-with: Cursor
